+++ R/subset_SingleGrainData.R
@@ -1,20 +1,21 @@
-#'@title Simple Subsetting of Single Grain Data from Risø BIN/BINX files
+#' @title Simple Subsetting of Single Grain Data from Risø BIN/BINX files
 #'
-#'@description Most measured single grains do not exhibit light and it makes
-#'usually sense to subset single grain datasets using a table of
-#'position and grain pairs
+#' @description
+#' Most measured single grains do not exhibit light and it usually makes sense
+#' to subset single grain datasets using a table of position and grain pairs.
 #'
-#'@param object [Luminescence::Risoe.BINfileData-class] (**required**): input object with the
-#'data to subset
+#' @param object [Luminescence::Risoe.BINfileData-class] (**required**):
+#' input object with the data to subset.
 #'
-#'@param selection [data.frame] (**required**): selection table with two columns
-#'for position (1st column) and grain (2nd column) (columns names do not matter)
+#' @param selection [data.frame], [matrix] (**required**):
+#' selection table with two columns for position and grain, respectively.
 #'
-#'@return A subset [Luminescence::Risoe.BINfileData-class] object
+#' @return A subset [Luminescence::Risoe.BINfileData-class] object
 #'
-#'@section Function version: 0.1.0
+#' @section Function version: 0.1.1
 #'
-#'@author Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
+#' @author
+#' Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
 #'
 #'@keywords manip datagen
 #'
@@ -42,6 +43,9 @@ subset_SingleGrainData <- function(
 
   ## Integrity checks -------------------------------------------------------
   .validate_class(object, "Risoe.BINfileData")
+  .validate_class(selection, c("data.frame", "matrix"))
+  if (ncol(selection) < 2)
+    .throw_error("'selection' should have 2 columns")
 
   ## try to work with selection
   selection <- as.data.frame(selection)[,1:2]
@@ -50,6 +54,8 @@ subset_SingleGrainData <- function(
 # Subset ------------------------------------------------------------------
   ## select ids for subsetting
   sel_id <-sort(merge(object@METADATA[,c("POSITION", "GRAIN", "ID")], selection)[["ID"]])
+  if (length(sel_id) == 0)
+    .throw_error("No matching records, check POSITION and GRAIN in 'selection'")
 
   ## pick data
   object@METADATA <- object@METADATA[sel_id,]